response to DF: cleanup
